Feat Personas Crud integration

+++ src/pages/home/botItem/botItem.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Bot } from "../../../data/types";
+import { Bot, Persona } from "../../../data/types";
 import botIcon from "../../../assets/bot-icon.png";
 import shareIcon from "../../../assets/share-icon.png";
 import { useAgents, AgentViewType } from "../../../context/botsContext";
@@ -11,8 +11,8 @@ import AssistantsService from "../../../services/assistants.service";
 import "./botItem.css";
 
 interface BotItemProps {
-  botItemData: Bot;
-  onBotSelect: (bot: Bot) => void;
+  botItemData: Bot | Persona;
+  onBotSelect: (bot: Bot | Persona) => void;
   onEditBot: (bot: Bot) => void;
   onBotDelete: (botId: string) => void;
 }
@@ -22,7 +22,6 @@ const BotItem: React.FC<BotItemProps> = ({ botItemData, onBotSelect, onEditBot, 
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [isShareModalOpen, setIsShareModalOpen] = useState(false);
 
-  // Asegúrate de que los IDs aquí coincidan exactamente con los proporcionados
   const personaImages: { [key: string]: string } = {
     "7bea4732-214f-40e7-9161-4e7241a2b97e": "https://static.vecteezy.com/system/resources/previews/026/536/284/non_2x/27yr-old-beautiful-face-ai-generated-free-photo.jpg",
     "7bea4732-214f-40e7-9161-4e7241a2b97f": "https://img.freepik.com/premium-photo/face-that-has-word-ai-it_872754-2069.jpg",
@@ -34,9 +33,11 @@ const BotItem: React.FC<BotItemProps> = ({ botItemData, onBotSelect, onEditBot, 
 
   const handleDelete = async () => {
     try {
-      const assistantsService = new AssistantsService();
-      await assistantsService.deleteAssistant(botItemData.id);
-      onBotDelete(botItemData.id);
+      if ('uri' in botItemData) {
+        const assistantsService = new AssistantsService();
+        await assistantsService.deleteAssistant(botItemData.id);
+        onBotDelete(botItemData.id);
+      }
       setIsModalOpen(false);
     } catch (error) {
       console.error("Error deleting assistant:", error);
@@ -52,14 +53,18 @@ const BotItem: React.FC<BotItemProps> = ({ botItemData, onBotSelect, onEditBot, 
   };
 
   const handleEditClick = () => {
-    onEditBot(botItemData);
+    if ('uri' in botItemData) {
+      onEditBot(botItemData);
+    }
   };
 
   const handleGoToCourseClick = () => {
     onBotSelect(botItemData);
   };
 
-  const imageUrl = personaImages[botItemData.persona_id || ""] || botIcon;
+  const imageUrl = 'persona_id' in botItemData
+    ? personaImages[botItemData.persona_id || ""] || botIcon
+    : personaImages[botItemData.face_id || ""] || botIcon;
 
   return (
     <div className="bot-card">
@@ -78,10 +83,10 @@ const BotItem: React.FC<BotItemProps> = ({ botItemData, onBotSelect, onEditBot, 
         <p className="bot-card-description">{botItemData.description}</p>
       </div>
       <div className="bot-buttons-area">
-        {agentViewType === AgentViewType.MyAgents ? (
+        {agentViewType === AgentViewType.MyAgents || AgentViewType.Personas ? (
           <>
             <SecondaryButton text="Delete" onClick={handleDeleteClick} enabled={true} />
-            <PrimaryButton text="Edit" onClick={handleEditClick} enabled={true} />
+            <PrimaryButton text="Edit" onClick={handleEditClick} enabled={!!('uri' in botItemData || AgentViewType.Personas)} />
           </>
         ) : (
           <PrimaryButton text="Go to Course" onClick={handleGoToCourseClick} enabled={true} />
